removing some empty and cancelled lines

+++ script.js
@@ -5,10 +5,14 @@ var start = function() {
 	grabNewUser.addEventListener('click', function(event) {
 		newUserPrompt();
 	})
-	/*grabRedButton.addEventListener('click', function(event) {
-		localStorage.clear();
-		location.reload();
-	})*/
+	grabRedButton.addEventListener('click', function(event) {
+		if (confirm('Are you sure you want to delete everything?')) {
+			localStorage.clear();
+			location.reload();
+		} else {
+			//Do Nothing.
+		}
+	})
 };
 
 var newUserPrompt = function() {
@@ -24,43 +28,6 @@ function userName() {
 	localStorage.clear();
 	var input = document.getElementById("userInput");
 	localStorage.setItem("$userName", input.value);
-	//localStorage.setItem("$category1", "rent");
-	//localStorage.setItem("$line1index0", 0);
-	//localStorage.setItem("$line1index1", 0);
-	//localStorage.setItem("$line1index2", 0);
-	/*localStorage.setItem("$category2", "phone");
-	localStorage.setItem("$category3", "food");
-	localStorage.setItem("$category4", "internet");
-	localStorage.setItem("$category5", "gifts");
-	localStorage.setItem("$category6", "vacation");
-	localStorage.setItem("$category7", "emergency");
-	var allCategories = [localStorage.getItem("$category1"), localStorage.getItem("$category2"), localStorage.getItem("$category3"), localStorage.getItem("$category4"), localStorage.getItem("$category5"), localStorage.getItem("$category6"), localStorage.getItem("$category7")];
-	
-
-	
-	localStorage.setItem("$line2index0", 0);
-	localStorage.setItem("$line3index0", 0);
-	localStorage.setItem("$line4index0", 0);
-	localStorage.setItem("$line5index0", 0);
-	localStorage.setItem("$line6index0", 0);
-	localStorage.setItem("$line7index0", 0);
-
-	
-	localStorage.setItem("$line2index1", 0);
-	localStorage.setItem("$line3index1", 0);
-	localStorage.setItem("$line4index1", 0);
-	localStorage.setItem("$line5index1", 0);
-	localStorage.setItem("$line6index1", 0);
-	localStorage.setItem("$line7index1", 0);
-
-	
-	localStorage.setItem("$line2index2", 0);
-	localStorage.setItem("$line3index2", 0);
-	localStorage.setItem("$line4index2", 0);
-	localStorage.setItem("$line5index2", 0);
-	localStorage.setItem("$line6index2", 0);
-	localStorage.setItem("$line7index2", 0);*/
-
 	localStorage.setItem("$myCurrentBalance", 0);
 	var allCategories = [];
 	localStorage.setItem("$allCategories", JSON.stringify(allCategories));
